pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cap"
version = "0.1.0"
description = "Counterfactual Audit Pipeline — GenAI-powered fairness auditing of facial analysis systems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ali Alenjani", email = "lenjani.a@gmail.com" }]
keywords = ["fairness", "auditing", "generative-ai", "facial-analysis", "counterfactual"]

dependencies = [
    # Core
    "numpy>=1.26,<2.0",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "click>=8.1",
    "rich>=13.7",
    "loguru>=0.7",

    # Diffusion stack (CUDA path; install torch separately for the right CUDA build)
    "diffusers>=0.30",
    "transformers>=4.44",
    "accelerate>=0.34",
    "safetensors>=0.4",
    "huggingface-hub>=0.24",
    "peft>=0.12",
    "bitsandbytes>=0.43",  # FP8/quantization on L4
    "sentencepiece>=0.2",
    "protobuf>=4.25",

    # Image / face
    "Pillow>=10.4",
    "opencv-python-headless>=4.10",
    "insightface>=0.7",  # ArcFace identity validation
    "onnxruntime-gpu>=1.18",  # InsightFace backend
    "deepface>=0.0.93",
    "facexlib>=0.3",  # face restoration / parsing utilities

    # Statistics
    "scipy>=1.13",
    "statsmodels>=0.14",
    "scikit-learn>=1.5",
    "pingouin>=0.5",  # convenient ANOVA / post-hoc / effect sizes

    # Visualization (publication + interactive)
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "plotly>=5.22",
    "altair>=5.3",
    "vegafusion>=1.6",  # for large-data Altair / Vega
    "vegafusion-python-embed>=1.6",
    "kaleido>=0.2",  # static export for plotly
    "colorcet>=3.1",  # perceptually uniform palettes
    "matplotlib-scalebar>=0.8",

    # Cloud / IO
    "google-cloud-storage>=2.18",
    "boto3>=1.34",  # AWS Rekognition
    "azure-ai-vision-imageanalysis>=1.0",
    "google-cloud-vision>=3.7",
    "requests>=2.32",
    "tenacity>=8.5",  # retry logic for flaky APIs
    "ratelimit>=2.2",  # rate limiting for APIs

    # Reproducibility
    "wandb>=0.17",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.6",
    "ruff>=0.5",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "jupyterlab>=4.2",
]

[project.scripts]
cap-generate = "cap.cli.generate:main"
cap-audit = "cap.cli.audit:main"
cap-analyze = "cap.cli.analyze:main"
cap-visualize = "cap.cli.visualize:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cap"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "gpu: requires CUDA GPU (skipped on CPU-only)",
    "slow: takes more than 10 seconds",
    "integration: hits real APIs (requires credentials)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RET"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
